+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - 2025-07-28
+- Create Stuff 'N Additions CF59 patch for fluid handler capabilities
+  - Extended fluid handler patches to all gadgets
+  - Read capacities from Create SA config file
+
 ## [0.4.0] - 2025-07-27
 
 ### Added